mocks renamed to fakedb
