+++ akka-projection-fakedb/src/main/scala/akka/projection/fakedb/DBIO.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2019 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package akka.projection.mocks
+package akka.projection.fakedb
 
 import scala.concurrent.Future
