--- akka-projection-mocks/src/main/scala/akka/projection/mocks/MockedDb.scala
@@ -1,9 +1,0 @@
-/*
- * Copyright (C) 2019 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package akka.projection.mocks
-
-class MockedDb {
-
-}
